extract parser functions out to helpers/parsers.js
